+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 - Support external addresses for balance fetching and requesting faucet funds.
+- Add enumerator pattern for address.transfers and address.trades.
 
 ## [0.0.8] - 2024-06-18